fixed Bug #2678 +  #2683: Inline default choice - highlight and Bug #2683: Qti Creator + Saving an item for the first time produces a PHP error in the logs
git-svn-id: http://vcs.taotesting.com/svn/oat/projects/taoQtiItem/trunk@20828 62dfeb86-a93f-4914-9d13-15c0a160985d

+++ helpers/CssHelper.php
@@ -63,7 +63,8 @@ EOF;
         $service = \taoItems_models_classes_ItemsService::singleton();
         $cssFile  = $service->getItemFolder($item, $lang) . $styleSheetPath;
 
-        if(empty($cssArr)) {
+        // make sure that 'no custom css' means exactly that
+        if(empty($cssArr) && file_exists($cssFile)) {
             unlink($cssFile);
         }
